refactor: move AskValidators and Directory extension
